Add OrchestratorHarness for in-process orchestrator tests
Extracts the orchestrator startup/lifecycle logic from serve.rs into a
new harness.rs library API. Tests can now hold Arc<AnyEventLog> directly
and use EventLog::subscribe() for event-driven waits instead of polling
SQLite files from subprocesses. Includes a proof integration test
(stream_trigger_route_uses_generic_stream_connector_in_process) and
fixes the OTEL SimpleSpanProcessor deadlock that occurs with
current-thread runtimes by using a multi-thread worker pool in the
harness background thread.
